requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0

SQLAlchemy==2.0.27
aiomysql==0.2.0

reportlab==4.1.0
qrcode==7.4.2
Pillow==10.2.0

huaweicloudsdkcore==3.1.176
huaweicloudsdkobs==3.1.176
esdk-obs-python==3.23.6